Use type-erased factory for valve
